requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.0